+++ workflows/offline_ml_diags/offline_ml_diags/compute_diags.py
@@ -55,8 +55,7 @@ def _write_nc(ds: xr.Dataset, output_dir: str, output_file: str):
 
     with NamedTemporaryFile() as tmpfile:
         ds.to_netcdf(tmpfile.name)
-        gsutil.copy(tmpfile.name, output_file)
-        # get_fs(output_dir).put(tmpfile.name, output_file)
+        get_fs(output_dir).put(tmpfile.name, output_file)
     logger.info(f"Writing netcdf to {output_file}")
 
 
@@ -69,7 +68,7 @@ if __name__ == "__main__":
         config = yaml.safe_load(f)
 
     logger.info("Reading grid...")
-    cat = intake.open_catalog("../../catalog.yml")
+    cat = intake.open_catalog("catalog.yml")
     grid = cat["grid/c48"].to_dask()
     grid = grid.drop(labels=["y_interface", "y", "x_interface", "x"])
     land_sea_mask = cat["landseamask/c48"].to_dask()